+++ CHANGELOG.md
@@ -25,6 +25,10 @@ release.
   ([#799](https://github.com/open-telemetry/opentelemetry-demo/pull/799))
 * Update frontend JavaScript SDKs to: 1.10.1/0.36.x
   ([#805](https://github.com/open-telemetry/opentelemetry-demo/pull/805))
+* Fix http.status_code on error in frontend
+  ([#810](https://github.com/open-telemetry/opentelemetry-demo/pull/810))
+* Fix bug in shipping calculation
+  ([#814](https://github.com/open-telemetry/opentelemetry-demo/pull/814))
 
 ## v0.1.0